requirements.txt
autopep8
black
flake8
isort
pandas
pyyaml
playwright
beautifulsoup4
go-task-bin
psycopg2-binary
yamllint
pre-commit
# nltk
# pre-commit install 
